Add files via upload
For this project it requires following  prerequisites=
1)Fundamental concepts of Machine Learning and Neural Network(preferably recurrent neural networks)
2)LSTM network working in Python and Keras

Named Entity Recognition using LSTMs with Keras. In this project,i used the Keras API with TensorFlow as its backend to build and train a bidirectional LSTM neural network model to recognize named entities in text data.

The task of NER is to find the type of words in the texts.Lets take example
let’s say you are working in the newspaper industry as an editor and you receive thousands of stories every day. How will you find the story which is related to specific sections like sports, politics, etc? Will you go through all of these stories? No, right?
How about a system that helps you segment into different categories? The system may also perform sophisticated tasks like separating stories city wise, identifying the person names involved in the story, organizations and so on.
This is where NER can be use for extraction of relevant information and categorizing them into predefine categories.